requirements.txt
adafruit-board-toolkit==1.1.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
astroid==3.2.4
asttokens==2.4.1
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.2.1
bitarray==2.9.3
bitstring==4.2.3
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.0
dill==0.3.9
disnake==2.10.1
docutils==0.21.2
ecdsa==0.19.0
esptool==4.8.1
frozenlist==1.5.0
idna==3.10
intelhex==2.3.0
isort==5.13.2
jedi==0.19.2
mccabe==0.7.0
multidict==6.2.0
mypy==1.11.2
mypy-extensions==1.0.0
paramiko==3.4.1
parso==0.8.4
platformdirs==4.3.6
propcache==0.3.1
pycparser==2.22
pylint==3.2.7
PyNaCl==1.5.0
pyserial==3.5
PyYAML==6.0.2
reedsolo==1.7.0
Send2Trash==1.8.3
six==1.17.0
thonny==4.1.7
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
websockets==13.0.1
yarl==1.19.0
